pyproject.toml
[project]
name = "yoonkyung-lee"
version = "0.1.0"
description = "Personal project with JupyterLab"
requires-python = ">=3.10"
dependencies = [
    "jupyterlab>=4.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
    "python-dotenv>=1.1.1",
    "pandas>=2.3.2",
    "ipython>=8.37.0",
    "termcolor>=3.1.0",
    "tiktoken>=0.11.0",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["week2", "week3"]

[tool.uv]
dev-dependencies = []
